IntegerValidator, Ipv4Validator, Ipv6Validator, NumericValidator, StartsWithValidator, StringValidator

+++ src/framework/validator/laravel/digits-between-validator.js
@@ -67,10 +67,10 @@ Object.assign(abv, function () {
             return [
                 {
                     'name': 'min',
-                    'type': 'integer'
+                    'type': 'numeric'
                 }, {
                     'name': 'max',
-                    'type': 'integer'
+                    'type': 'numeric'
                 }
             ];
         }
